[#9056] Alert definition groups implementation + APIs for groups and templates
Summary:
Implemented Alert Definition Groups as described in PRD and UI mocks.
Groups manages underlying definitions for universes.
Also, group is linked to alert route.

For now we're creating some groups automatically on Customer creation for all universes.
Others are managed manually through API. Automatic ones can also be managed after initial creation through API.

Also added swagger annotations to all the AlertController methods

Test Plan:
1. Install old version of the platform.
2. Create 2 universes.
3. Raise memory consumption alert by setting threshold to 2.
4. Create replication lag alert definitions -one active and one inactive.
5. Upgrade platform to the version with these changes.
6. Make sure all alert definitions are in place both in DB and on disk. Make sure 5 alert definition groups are in DB - Memory Consumption, Clock Skew and 2 Replication Lag for both universes - active and inactive.
7. Make Memory Consumption Alert resolved. Make sure email is received on default SMTP receiver.
8. Create new universe. Make sure alert definitions for it are created both in DB and on disk.
9. Create replication lag alert definition group for new universe via UI.
10. List, Create, Update, Get and Delete alert definition groups via API. Also query alert definition templates via API.
11. Via API configure custom SMTP receiver and route. Assign it to Memory consumption alert definition.
12. Raise and clean memory consumption alert. Make sure notifications are received on custom receiver.
13. Delete all universes. Make sure add definition based alerts are resolved + no alert definitions in DB related to deleted universes + no replication lag groups, related to deleted universes.

Reviewers: spotachev

Reviewed By: spotachev

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D12124

+++ managed/src/main/java/com/yugabyte/yw/controllers/AlertController.java
@@ -18,116 +18,196 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.google.inject.Inject;
+import com.yugabyte.yw.common.AlertDefinitionTemplate;
 import com.yugabyte.yw.common.YWServiceException;
 import com.yugabyte.yw.common.alerts.*;
-import com.yugabyte.yw.forms.AlertDefinitionFormData;
 import com.yugabyte.yw.forms.AlertReceiverFormData;
 import com.yugabyte.yw.forms.AlertRouteFormData;
+import com.yugabyte.yw.forms.YWResults;
 import com.yugabyte.yw.models.*;
-import com.yugabyte.yw.models.filters.AlertDefinitionFilter;
+import com.yugabyte.yw.models.filters.AlertDefinitionGroupFilter;
+import com.yugabyte.yw.models.filters.AlertDefinitionTemplateFilter;
 import com.yugabyte.yw.models.filters.AlertFilter;
-import com.yugabyte.yw.models.helpers.KnownAlertLabels;
+import com.yugabyte.yw.models.paging.AlertDefinitionGroupPagedQuery;
+import com.yugabyte.yw.models.paging.AlertDefinitionGroupPagedResponse;
+import io.swagger.annotations.*;
 import lombok.extern.slf4j.Slf4j;
-import play.data.Form;
 import play.libs.Json;
 import play.mvc.Result;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Slf4j
+@Api(value = "Alert", authorizations = @Authorization(AbstractPlatformController.API_KEY_AUTH))
 public class AlertController extends AuthenticatedController {
+
+  @Inject private AlertDefinitionGroupService alertDefinitionGroupService;
 
   @Inject private AlertDefinitionService alertDefinitionService;
 
   @Inject private AlertService alertService;
 
   /** Lists alerts for given customer. */
+  @ApiOperation(value = "listAlerts", response = Alert.class, responseContainer = "List")
   public Result list(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
 
     ArrayNode alerts = Json.newArray();
     AlertFilter filter = AlertFilter.builder().customerUuid(customerUUID).build();
     alertService.list(filter).forEach(alert -> alerts.add(Json.toJson(alert)));
-    return ok(alerts);
+    return YWResults.withData(alerts);
   }
 
+  @ApiOperation(value = "listActiveAlerts", response = Alert.class, responseContainer = "List")
   public Result listActive(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
 
     ArrayNode alerts = Json.newArray();
     AlertFilter filter = AlertFilter.builder().customerUuid(customerUUID).build();
     alertService.listNotResolved(filter).forEach(alert -> alerts.add(Json.toJson(alert)));
-    return ok(alerts);
+    return YWResults.withData(alerts);
   }
 
-  public Result createDefinition(UUID customerUUID, UUID universeUUID) {
-
+  @ApiOperation(value = "getDefinitionGroup", response = AlertDefinitionGroup.class)
+  public Result getDefinitionGroup(UUID customerUUID, UUID groupUUID) {
     Customer.getOrBadRequest(customerUUID);
 
-    Form<AlertDefinitionFormData> formData =
-        formFactory.getFormDataOrBadRequest(AlertDefinitionFormData.class);
+    AlertDefinitionGroup group = alertDefinitionGroupService.getOrBadRequest(groupUUID);
 
-    AlertDefinitionFormData data = formData.get();
-    Universe universe = Universe.getOrBadRequest(universeUUID);
-
-    AlertDefinition definition =
-        new AlertDefinition()
-            .setCustomerUUID(customerUUID)
-            .setName(data.name)
-            .setQuery(data.template.buildTemplate(universe.getUniverseDetails().nodePrefix))
-            .setQueryThreshold(data.value)
-            .setActive(true)
-            .setLabels(AlertDefinitionLabelsBuilder.create().appendTarget(universe).get());
-    AlertDefinition createdDefinition = alertDefinitionService.create(definition);
-
-    return ok(Json.toJson(createdDefinition));
+    return YWResults.withData(group);
   }
 
-  public Result getAlertDefinition(UUID customerUUID, UUID universeUUID, String name) {
-    Customer.getOrBadRequest(customerUUID);
+  @ApiOperation(
+      value = "listDefinitionGroupTemplates",
+      response = AlertDefinitionGroup.class,
+      responseContainer = "List")
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "ListTemplatesRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.models.filters.AlertDefinitionTemplateFilter",
+          required = true))
+  public Result listDefinitionGroupTemplates(UUID customerUUID) {
+    Customer customer = Customer.getOrBadRequest(customerUUID);
 
-    List<AlertDefinition> definitions =
-        alertDefinitionService.list(
-            AlertDefinitionFilter.builder()
-                .customerUuid(customerUUID)
-                .name(name)
-                .label(KnownAlertLabels.UNIVERSE_UUID, universeUUID.toString())
-                .build());
-    AlertDefinition definition =
-        definitions
-            .stream()
-            .findFirst()
-            .orElseThrow(
-                () ->
-                    new YWServiceException(
-                        BAD_REQUEST,
-                        name
-                            + " alert definition for customer "
-                            + customerUUID.toString()
-                            + " and universe "
-                            + universeUUID
-                            + " not found"));
+    AlertDefinitionTemplateFilter filter =
+        Json.fromJson(request().body().asJson(), AlertDefinitionTemplateFilter.class);
 
-    return ok(Json.toJson(definition));
+    List<AlertDefinitionGroup> groups =
+        Arrays.stream(AlertDefinitionTemplate.values())
+            .filter(filter::matches)
+            .map(
+                template -> alertDefinitionGroupService.createGroupFromTemplate(customer, template))
+            .collect(Collectors.toList());
+
+    return YWResults.withData(groups);
   }
 
-  public Result updateAlertDefinition(UUID customerUUID, UUID alertDefinitionUUID) {
-
+  @ApiOperation(value = "pageDefinitionGroups", response = AlertDefinitionGroupPagedResponse.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "PageDefinitionGroupsRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.models.paging.AlertDefinitionGroupPagedQuery",
+          required = true))
+  public Result pageDefinitionGroups(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
 
-    AlertDefinition definition = alertDefinitionService.getOrBadRequest(alertDefinitionUUID);
+    AlertDefinitionGroupPagedQuery query =
+        Json.fromJson(request().body().asJson(), AlertDefinitionGroupPagedQuery.class);
+    query.setFilter(query.getFilter().toBuilder().customerUuid(customerUUID).build());
 
-    AlertDefinitionFormData data =
-        formFactory.getFormDataOrBadRequest(AlertDefinitionFormData.class).get();
+    AlertDefinitionGroupPagedResponse groups = alertDefinitionGroupService.pagedList(query);
 
-    // For now only templates related to universe are supported.
-    Universe universe = Universe.getOrBadRequest(definition.getUniverseUUID());
-    definition.setQuery(data.template.buildTemplate(universe.getUniverseDetails().nodePrefix));
-    definition.setQueryThreshold(data.value);
-    definition.setActive(data.active);
-    AlertDefinition updatedDefinition = alertDefinitionService.update(definition);
-    return ok(Json.toJson(updatedDefinition));
+    return YWResults.withData(groups);
+  }
+
+  @ApiOperation(
+      value = "listDefinitionGroups",
+      response = AlertDefinitionGroup.class,
+      responseContainer = "List")
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "ListGroupsRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.models.filters.AlertDefinitionGroupFilter",
+          required = true))
+  public Result listDefinitionGroups(UUID customerUUID) {
+    Customer.getOrBadRequest(customerUUID);
+
+    AlertDefinitionGroupFilter filter =
+        Json.fromJson(request().body().asJson(), AlertDefinitionGroupFilter.class);
+
+    List<AlertDefinitionGroup> groups = alertDefinitionGroupService.list(filter);
+
+    return YWResults.withData(groups);
+  }
+
+  @ApiOperation(value = "createDefinitionGroup", response = AlertDefinitionGroup.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "CreateGroupRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.models.AlertDefinitionGroup",
+          required = true))
+  public Result createDefinitionGroup(UUID customerUUID) {
+    Customer.getOrBadRequest(customerUUID);
+
+    AlertDefinitionGroup group =
+        Json.fromJson(request().body().asJson(), AlertDefinitionGroup.class);
+
+    if (group.getUuid() != null) {
+      throw new YWServiceException(BAD_REQUEST, "Can't create group with uuid set");
+    }
+
+    group = alertDefinitionGroupService.save(group);
+
+    auditService().createAuditEntry(ctx(), request());
+    return YWResults.withData(group);
+  }
+
+  @ApiOperation(value = "updateDefinitionGroup", response = AlertDefinitionGroup.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "UpdateGroupRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.models.AlertDefinitionGroup",
+          required = true))
+  public Result updateDefinitionGroup(UUID customerUUID, UUID groupUUID) {
+    Customer.getOrBadRequest(customerUUID);
+    alertDefinitionGroupService.getOrBadRequest(groupUUID);
+
+    AlertDefinitionGroup group =
+        Json.fromJson(request().body().asJson(), AlertDefinitionGroup.class);
+
+    if (group.getUuid() == null) {
+      throw new YWServiceException(BAD_REQUEST, "Can't update group with missing uuid");
+    }
+
+    if (!group.getUuid().equals(groupUUID)) {
+      throw new YWServiceException(
+          BAD_REQUEST, "Group UUID from path should be consistent with body");
+    }
+
+    group = alertDefinitionGroupService.save(group);
+
+    auditService().createAuditEntry(ctx(), request());
+    return YWResults.withData(group);
+  }
+
+  @ApiOperation(value = "deleteDefinitionGroup", response = YWResults.YWSuccess.class)
+  public Result deleteDefinitionGroup(UUID customerUUID, UUID groupUUID) {
+    Customer.getOrBadRequest(customerUUID);
+
+    alertDefinitionGroupService.getOrBadRequest(groupUUID);
+
+    alertDefinitionGroupService.delete(groupUUID);
+
+    auditService().createAuditEntry(ctx(), request());
+    return YWResults.YWSuccess.empty();
   }
 
   /**
@@ -144,6 +224,13 @@ public class AlertController extends AuthenticatedController {
     }
   }
 
+  @ApiOperation(value = "createAlertReceiver", response = AlertReceiver.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "CreateAlertReceiverRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.forms.AlertReceiverFormData",
+          required = true))
   public Result createAlertReceiver(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
 
@@ -163,14 +250,22 @@ public class AlertController extends AuthenticatedController {
 
     receiver.save();
     auditService().createAuditEntryWithReqBody(ctx());
-    return ok(Json.toJson(receiver));
+    return YWResults.withData(receiver);
   }
 
+  @ApiOperation(value = "getAlertReceiver", response = AlertReceiver.class)
   public Result getAlertReceiver(UUID customerUUID, UUID alertReceiverUUID) {
     Customer.getOrBadRequest(customerUUID);
-    return ok(Json.toJson(AlertReceiver.getOrBadRequest(customerUUID, alertReceiverUUID)));
+    return YWResults.withData(AlertReceiver.getOrBadRequest(customerUUID, alertReceiverUUID));
   }
 
+  @ApiOperation(value = "updateAlertReceiver", response = AlertReceiver.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "UpdateAlertReceiverRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.forms.AlertReceiverFormData",
+          required = true))
   public Result updateAlertReceiver(UUID customerUUID, UUID alertReceiverUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertReceiver receiver = AlertReceiver.getOrBadRequest(customerUUID, alertReceiverUUID);
@@ -188,9 +283,10 @@ public class AlertController extends AuthenticatedController {
 
     receiver.save();
     auditService().createAuditEntryWithReqBody(ctx());
-    return ok(Json.toJson(receiver));
+    return YWResults.withData(receiver);
   }
 
+  @ApiOperation(value = "deleteAlertReceiver", response = YWResults.YWSuccess.class)
   public Result deleteAlertReceiver(UUID customerUUID, UUID alertReceiverUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertReceiver receiver = AlertReceiver.getOrBadRequest(customerUUID, alertReceiverUUID);
@@ -202,14 +298,25 @@ public class AlertController extends AuthenticatedController {
     }
 
     auditService().createAuditEntry(ctx(), request());
-    return ok();
+    return YWResults.YWSuccess.empty();
   }
 
+  @ApiOperation(
+      value = "listAlertReceivers",
+      response = AlertReceiver.class,
+      responseContainer = "List")
   public Result listAlertReceivers(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
-    return ok(Json.toJson(AlertReceiver.list(customerUUID)));
+    return YWResults.withData(AlertReceiver.list(customerUUID));
   }
 
+  @ApiOperation(value = "createAlertRoute", response = AlertRoute.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "CreateAlertRouteRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.forms.AlertRouteFormData",
+          required = true))
   public Result createAlertRoute(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertRouteFormData data = formFactory.getFormDataOrBadRequest(AlertRouteFormData.class).get();
@@ -217,17 +324,25 @@ public class AlertController extends AuthenticatedController {
     try {
       AlertRoute route = AlertRoute.create(customerUUID, data.name, receivers);
       auditService().createAuditEntryWithReqBody(ctx());
-      return ok(Json.toJson(route));
+      return YWResults.withData(route);
     } catch (Exception e) {
       throw new YWServiceException(BAD_REQUEST, "Unable to create alert route: " + e.getMessage());
     }
   }
 
+  @ApiOperation(value = "getAlertRoute", response = AlertRoute.class)
   public Result getAlertRoute(UUID customerUUID, UUID alertRouteUUID) {
     Customer.getOrBadRequest(customerUUID);
-    return ok(Json.toJson(AlertRoute.getOrBadRequest(customerUUID, alertRouteUUID)));
+    return YWResults.withData(AlertRoute.getOrBadRequest(customerUUID, alertRouteUUID));
   }
 
+  @ApiOperation(value = "updateAlertRoute", response = AlertRoute.class)
+  @ApiImplicitParams(
+      @ApiImplicitParam(
+          name = "UpdateAlertRouteRequest",
+          paramType = "body",
+          dataType = "com.yugabyte.yw.forms.AlertRouteFormData",
+          required = true))
   public Result updateAlertRoute(UUID customerUUID, UUID alertRouteUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertRouteFormData data = formFactory.getFormDataOrBadRequest(AlertRouteFormData.class).get();
@@ -239,28 +354,47 @@ public class AlertController extends AuthenticatedController {
       route.save();
 
       auditService().createAuditEntryWithReqBody(ctx());
-      return ok(Json.toJson(route));
+      return YWResults.withData(route);
     } catch (Exception e) {
       throw new YWServiceException(BAD_REQUEST, "Unable to update alert route: " + e.getMessage());
     }
   }
 
+  @ApiOperation(value = "deleteAlertRoute", response = YWResults.YWSuccess.class)
   public Result deleteAlertRoute(UUID customerUUID, UUID alertRouteUUID) {
     Customer.getOrBadRequest(customerUUID);
     AlertRoute route = AlertRoute.getOrBadRequest(customerUUID, alertRouteUUID);
     log.info("Deleting alert route {} for customer {}", alertRouteUUID, customerUUID);
 
+    AlertDefinitionGroupFilter groupFilter =
+        AlertDefinitionGroupFilter.builder().routeUuid(route.getUuid()).build();
+    List<AlertDefinitionGroup> groups = alertDefinitionGroupService.list(groupFilter);
+    if (!groups.isEmpty()) {
+      throw new YWServiceException(
+          BAD_REQUEST,
+          "Unable to delete alert route: "
+              + alertRouteUUID
+              + ". "
+              + groups.size()
+              + " alert definition groups are linked to it. Examples: "
+              + groups
+                  .stream()
+                  .limit(5)
+                  .map(AlertDefinitionGroup::getName)
+                  .collect(Collectors.toList()));
+    }
     if (!route.delete()) {
       throw new YWServiceException(
           INTERNAL_SERVER_ERROR, "Unable to delete alert route: " + alertRouteUUID);
     }
 
     auditService().createAuditEntry(ctx(), request());
-    return ok();
+    return YWResults.YWSuccess.empty();
   }
 
+  @ApiOperation(value = "listAlertRoutes", response = AlertRoute.class, responseContainer = "List")
   public Result listAlertRoutes(UUID customerUUID) {
     Customer.getOrBadRequest(customerUUID);
-    return ok(Json.toJson(AlertRoute.listByCustomer(customerUUID)));
+    return YWResults.withData(AlertRoute.listByCustomer(customerUUID));
   }
 }
